// Front/src/enums/clickTypes.ts
export const enum ClickTypes {
    profile = "profile",
    more = "more",
    moreProfile = "moreProfile",
    comments = "comments",
    send = "send",
    like = "like",
    likeComment = "likeComment",
    save = "save",
    notifications = "notifications",
    messages = "messages",
    story = "story",
    settings = "settings",
    posts = "posts",
    back = "back",
    backToProfile = "backToProfile",
    backToHome = "backToHome",
    backToChat = "backToChat",
    archiveBack = "archiveBack",
    aboutBack = "aboutBack",
    backToSettings = "backToSettings",
    saved = "saved",
    chats = "chats",
    chat = "chat",
    logout = "logout",
    archive = "archive",
    blocked = "blocked",
    followers = "followers",
    following = "following",
}